refactor(releaserc): Update releaserc file paths
This commit updates the paths for the releaserc files in the semantic-release.sh script. The pre-release configuration file path is changed to ".releaserc.pre-release.js" and the release configuration file path is changed to ".releaserc.release.js". This ensures that the correct configuration file is used based on the release type.

Signed-off-by: David Hooton <david@hooton.org>